_legacy-v1-python/requirements.txt
# AdCraft Backend — Python 3.12
# Versões fixadas para builds reproduzíveis.
# Atualizar com: pip-compile requirements.in

# ---------------------------------------------------------------------------
# Web framework
# ---------------------------------------------------------------------------
fastapi==0.110.0
uvicorn[standard]==0.27.1     # inclui websockets, httptools, watchfiles

# ---------------------------------------------------------------------------
# Task queue
# ---------------------------------------------------------------------------
celery[redis]==5.3.6
redis==5.0.3                  # cliente Python para Redis (broker + resultados)

# ---------------------------------------------------------------------------
# AI — Google Gemini SDK
# ---------------------------------------------------------------------------
google-genai==1.1.0

# ---------------------------------------------------------------------------
# Banco de dados — Supabase
# ---------------------------------------------------------------------------
supabase==2.4.3

# ---------------------------------------------------------------------------
# Storage — Cloudflare R2 (S3-compatible via boto3)
# ---------------------------------------------------------------------------
boto3==1.34.69
botocore==1.34.69             # boto3 depende de botocore; pin conjunto

# ---------------------------------------------------------------------------
# Segurança — criptografia de credenciais (AES-256 via Fernet)
# ---------------------------------------------------------------------------
cryptography==42.0.5

# ---------------------------------------------------------------------------
# Validação e configuração
# ---------------------------------------------------------------------------
pydantic>=2.7.0
pydantic-settings>=2.2.0

# ---------------------------------------------------------------------------
# HTTP client assíncrono (ferramentas de busca, Ad Library, YouTube)
# ---------------------------------------------------------------------------
httpx==0.27.0

# ---------------------------------------------------------------------------
# Upload de arquivos multipart (FastAPI)
# ---------------------------------------------------------------------------
python-multipart==0.0.9

# ---------------------------------------------------------------------------
# Processamento de vídeo — FFmpeg via subprocess Python
# ---------------------------------------------------------------------------
ffmpeg-python==0.2.0

# ---------------------------------------------------------------------------
# Testes
# ---------------------------------------------------------------------------
pytest==8.1.1
pytest-asyncio==0.23.6        # asyncio_mode = auto no pytest.ini
pytest-cov==5.0.0
pytest-xdist==3.5.0           # execução paralela (-n auto)
